backend/requirements.txt
fastapi
uvicorn
pandas
scikit-learn
openpyxl
python-multipart
python-dotenv
google-generativeai
